@@ -1,42 +1,53 @@
 {
-  "name": "vite-plugin-inline-css-modules",
-  "version": "0.0.7",
+  "name": "@cueaz/vite-plugin-inline-css-modules",
+  "version": "1.0.3",
   "description": "Write CSS modules without leaving your javascript!",
   "author": "Mia Korennykh",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/bluskript/vite-plugin-inline-css-modules.git"
+    "url": "git+https://github.com/cueaz/vite-plugin-inline-css-modules.git"
   },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
-    }
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "require": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
+    },
+    "./global": {
+      "types": "./global.d.ts"
+    },
+    "./package.json": "./package.json"
   },
   "files": [
-    "dist"
+    "dist",
+    "global.d.ts"
   ],
   "scripts": {
-    "dev": "pnpm build -- --watch",
-    "build": "tsup src/index.ts",
-    "build:prod": "cross-env NODE_ENV=production pnpm build"
+    "build": "tsup",
+    "format": "prettier --write ."
+  },
+  "dependencies": {
+    "magic-string": "^0.30.8"
   },
   "peerDependencies": {
     "vite": ">2.0.0-0"
   },
   "devDependencies": {
-    "@alloc/prettier-config": "^1.0.0",
-    "@types/node": "^14.14.20",
-    "cross-env": "^7.0.3",
-    "prettier": "^2.0.5",
-    "tsup": "^5.12.6",
-    "typescript": "4.8.0-dev.20220806",
-    "vite": "latest"
+    "@types/node": "^20.11.25",
+    "prettier": "^3.2.5",
+    "tsup": "^8.0.2",
+    "typescript": "5.4.2",
+    "vite": "^5.1.5"
   },
-  "prettier": "@alloc/prettier-config",
   "keywords": [
     "vite",
     "vite-plugin"
