@@ -1,6 +1,7 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { Root } from './Root'
-import './styles.css'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { Root } from './Root';
+import './styles.css';
 
-ReactDOM.render(<Root />, document.getElementById('root'))
+const root = ReactDOM.createRoot(document.getElementById('root')!);
+root.render(<Root />);
