@@ -1,10 +1,9 @@
-import React from 'react'
-import { css } from 'vite-plugin-inline-css-modules'
+import React from 'react';
 
 const classes = css`
   .root {
     background-color: #1f1e33;
   }
-`
+`;
 
-export const Root = () => <div className={classes.root}>Hello world</div>
+export const Root = () => <div className={classes.root}>Hello world</div>;
