@@ -1,15 +1,10 @@
-import { resolve } from 'path'
-import { env } from 'process'
-import type { Options } from 'tsup'
+import { defineConfig } from 'tsup';
 
-const isProd = env.NODE_ENV === 'production'
-
-export const tsup: Options = {
+export default defineConfig({
+  entry: ['src/index.ts'],
   clean: true,
   dts: true,
-  entryPoints: [resolve(__dirname, 'src', 'index.ts')],
-  minify: isProd,
   sourcemap: true,
   splitting: false,
   format: ['cjs', 'esm'],
-}
+});
